+++ get_x_navigation_generator/lib/src/generators/get_x_navigation_config_generator.dart
@@ -6,6 +6,7 @@ import 'package:code_builder/code_builder.dart';
 import 'package:dart_style/dart_style.dart';
 import 'package:get_x_navigation_generator/src/code_builder/library_builder.dart';
 import 'package:get_x_navigation_generator/src/models/get_x_route_config.dart';
+import 'package:get_x_navigation_generator/src/resolvers/importable_type_resolver.dart';
 import 'package:get_x_navigation_generator_interface/get_x_navigation_generator_interface.dart';
 import 'package:glob/glob.dart';
 import 'package:source_gen/source_gen.dart';
@@ -17,11 +18,17 @@ class GetXNavigationConfigGenerator
   @override
   dynamic generateForAnnotatedElement(
       Element element, ConstantReader annotation, BuildStep buildStep) async {
+    final typeResolver =
+        ImportableTypeResolverImpl(await buildStep.resolver.libraries.toList());
     final configFiles = Glob("**.get_x_navigation.json");
     final navigatorClassName =
         annotation.peek('navigatorClassName')?.stringValue;
-
+    final pageTypeAsDartType = annotation.peek('pageType')?.typeValue;
+    final pageType = pageTypeAsDartType == null
+        ? null
+        : typeResolver.resolveType(pageTypeAsDartType);
     final jsonData = <Map>[];
+
     await for (final id in buildStep.findAssets(configFiles)) {
       final dynamic json = jsonDecode(await buildStep.readAsString(id));
       jsonData.addAll((json as List).map((dynamic e) => e as Map).toList());
@@ -36,6 +43,7 @@ class GetXNavigationConfigGenerator
       routes: routes,
       className: navigatorClassName ?? _navigatorClassNameDefault,
       targetFile: element.source?.uri,
+      pageType: pageType,
     );
 
     final generatedLib = generator.generate();